src/main/webapp/app/entities/peer/delete/peer-delete-dialog.component.ts
import { Component } from '@angular/core';
import { FormsModule } from '@angular/forms';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import SharedModule from 'app/shared/shared.module';
import { IPeer } from '../peer.model';
import { PeerService } from '../service/peer.service';
import { ITEM_DELETED_EVENT } from 'app/config/navigation.constants';

@Component({
  standalone: true,
  templateUrl: './peer-delete-dialog.component.html',
  imports: [SharedModule, FormsModule],
})
export class PeerDeleteDialogComponent {
  peer?: IPeer;

  constructor(protected peerService: PeerService, protected activeModal: NgbActiveModal) {}

  cancel(): void {
    this.activeModal.dismiss();
  }

  confirmDelete(id: number): void {
    this.peerService.delete(id).subscribe(() => {
      this.activeModal.close(ITEM_DELETED_EVENT);
    });
  }
}
